check expected turbostat residency for both s2i and longidle
the tool is turbostat which HWE suggested[1] and also provide package_state_show for debug reference.

This test job is an automatic test case and should be included into projects those need e-star compliance.

This test job was verified for mounths on pc-sanity [2], but is refactored according to suggestion of reviewer [3]

[1] https://trello.com/c/3C0b5Nkt
[2] https://git.launchpad.net/plainbox-provider-pc-sanity/tree/usr/share/plainbox-provider-checkbox/units/pc-sanity/pc-sanity-power-consumption.pxu#n27
[3] https://chat.canonical.com/canonical/pl/u7upmtk9stdgbbz7x7e4bbsw7o